* Changed the headers to point to the files in the sbpl_collision_checking package * changed a short unsigned int to int in the environment to work with the newer version of the occupancy grid

+++ sbpl_full_body_planning/sbpl_full_body_planner/src/environment_dualrobarm3d.cpp
@@ -1953,7 +1953,7 @@ void EnvironmentDUALROBARM3D::updateOccupancyGridFromCollisionMap(const arm_navi
 
 std::vector<std::vector<double> > EnvironmentDUALROBARM3D::getShortestPath()
 {
-  std::vector<short unsigned int> start(3,0);
+  std::vector<int> start(3,0);
   std::vector<double> waypoint(3,0);
   std::vector<std::vector<int> > path;
   std::vector<std::vector<double> > dpath;
